telegram-bot/requirements.txt
python-telegram-bot
aiohttp
asyncio
requests
python-telegram-bot-pagination